Rename "modules" webpack to "script-modules"
